+++ yamibolib/src/main/java/com/yamibo/main/yamibolib/model/UserProfile.java
@@ -131,4 +131,6 @@ public class UserProfile {
     public void setNewmypost(int newmypost) {
         this.newmypost = newmypost;
     }
+
+
 }
